refactor: SubmissionResult -> Submission 폴더로 옮김
