+++ frontend/src/components/Submission/ResultInfo.tsx
