+++ frontend/src/components/Submission/ResultInfoWrapper.tsx
