+++ frontend/src/components/Submission/ResultList.tsx
