+++ frontend/src/components/Submission/ResultTotalInfo.tsx
