+++ frontend/src/components/Submission/index.tsx
@@ -1,0 +1,1 @@
+export { SubmissionResult } from './SubmissionResult';
